Move Dasher to app.py. Update / add docstrings.
