+++ dasher/base.py
@@ -1,11 +1,37 @@
 from abc import ABC, abstractmethod
 from collections import OrderedDict
 
-# from dash.development.base_component import Component
 
+class BaseWidget(ABC):
+    """ Abstract base class of a dasher widget.
+    A dasher widget is an interactive control, which consists of an interactive dash
+    `component`, a `label` and a final `layout`.
 
-class DasherWidget(ABC):
+    Attributes
+    ----------
+    name: str
+        Name of the widget.
+    x: object
+        The object, whose type determines the type of the widget.
+    component: DasherComponent
+        The interactive dash component.
+    label: str
+        The label for the dash component.
+    layout: dash.development.base_component.Component
+        The `layout` is a styled and labeled version of `component`.
+    dependency: str, optional
+        The attribute used for the ``dash.Input`` dependency. Default: "value".
+    """
     def __init__(self, name, x, label=None, dependecy="value"):
+        """
+
+        Parameters
+        ----------
+        name
+        x
+        label
+        dependecy
+        """
         self.name = name
         self.x = x
         self.label = label if label is not None else name
@@ -28,7 +54,7 @@ class DasherWidget(ABC):
     @abstractmethod
     def layout(self):
         """ Abstract property. The implementation of the getter method in the child
-        class must return the layout of the concrete component.
+        class must return the final layout of the widget.
 
         Returns
         -------
@@ -38,8 +64,8 @@ class DasherWidget(ABC):
         pass
 
 
-class DasherWidgetPassthroughMixin(DasherWidget, ABC):
-    """ Passthrough mixin for custom dash components. """
+class WidgetPassthroughMixin(BaseWidget, ABC):
+    """ Passthrough mixin to support custom dash components. """
 
     @property
     def component(self):
@@ -50,100 +76,21 @@ class DasherWidgetPassthroughMixin(DasherWidget, ABC):
         return self.x
 
 
-class DasherComponent(ABC):
-    """ Abstract base class of dasher components.
-
-    Attributes
-    ----------
-    name: str
-        Name of the component.
-    x: object of supported type
-        Object, which was used to determine which interactive component was generated.
-    dependency: str
-        Property to use for the input dependency.
-    layout
-    """
-
-    def __init__(self, name, x, dependency="value"):
-        """
-        Parameters
-        ----------
-        name: str
-            Name of the component.
-        x: object of supported type
-            Object used to determine which interactive component is returned.
-        dependency: str, optional
-            Property to use for the input dependency.
-        """
-        self.name = name
-        self.x = x
-        self.dependency = dependency
-
-    @property
-    @abstractmethod
-    def layout(self):
-        """ Abstract property. The implementation of the getter method in the child
-        class must return the layout of the concrete component.
-
-        Returns
-        -------
-        dash.development.base_component.Component
-            Generated dash component.
-        """
-        pass
-
-
-# class DasherWidget(object):
-#     """ A dasher widget contains an interactive dasher component, a label and a layout.
-#     The layout combines the label and the interactive component into the final form,
-#     which is displayed in the app. The layout is generated by a subclass of
-#     ``DasherLayout``.
-#
-#     Attributes
-#     ----------
-#     component: DasherComponent
-#         An interactive dasher component.
-#     label: str
-#         A label for the interactive dasher component.
-#     layout: dash.development.base_component.Component
-#         `layout` is a dash component, which is diplayed in the dasher app.
-#     """
-#
-#     def __init__(self, component, label, layout):
-#         """
-#         Parameters
-#         ----------
-#         component: DasherComponent
-#             An interactive dasher component.
-#         label: str
-#             A label for the interactive dasher component.
-#         layout: dash.development.base_component.Component
-#             `layout` is a dash component, which is diplayed in the dasher app.
-#         """
-#         self.component = component
-#         self.label = label
-#         if not isinstance(layout, Component):
-#             raise TypeError("widget must be a dash Component")
-#         self.layout = layout
-
-
-class DasherLayout(ABC):
+class BaseLayout(ABC):
     """ Abtract base class of a dasher layout, which is responsible for creating the
     layout of the dasher app.
 
-    It renders dasher components into dasher widgets, thereby
-    combining the interactive component with a label and creating the final layout which
-    is displayed in the app (``render_component``).
+    The widget specification (`widget_spec`) is used to determine the types and the
+    layout of the automatically generated interactive widgets.
 
-    It handles the addition of callbacks and creates a suitable form of separation
-    between the callbacks in the app layout. The standard way to do this is to generate
-    a separate tab for each callback (``add_callback``).
+    A layout class also handles the addition of callbacks via the `add_callback` method
+    and creates a suitable form of separation between the callbacks in the app layout.
+    The standard way to do this is to generate a separate tab for each callback.
 
-    A child class must implement the abtract methods (``render_component`` and
-    ``add_callback``) and create the final app layout as its' ``layout`` attribute.
-    If the layout needs external stylesheets, the child class must announce this by
-    creating an ``external_stylesheets`` attribute containing the list of required
-    external stylesheets.
+    A child class must implement the abtract `add_callback` method and create the final
+    app layout as its' ``layout`` attribute. If the layout needs external stylesheets,
+    the child class must announce this by creating an `external_stylesheets` attribute
+    containing the list of required external stylesheets.
 
     Attributes
     ----------
@@ -152,7 +99,7 @@ class DasherLayout(ABC):
     credits: bool
         If true, dasher / layout credits are shown in the app.
     layout: list of dash.development.base_component.Component
-        The final app layout (is assigned to the ``layout`` property of the dash app)
+        The final app layout (is assigned to the ``layout`` property of the dash app).
     """
 
     output_base = "dasher-output"
@@ -163,6 +110,8 @@ class DasherLayout(ABC):
         ----------
         title: str
             Title of the dash app.
+        widget_spec: OrderedDict
+            Widget specification used to determine the types of the interactive widgets.
         credits: bool
             If true, dasher / layout credits are shown in the app.
         """
@@ -187,8 +136,8 @@ class DasherLayout(ABC):
 
         Parameters
         ----------
-        callback: DasherCallback
-            The dasher callback to add to the layout.
+        callback: Callback
+            The callback to add to the layout.
         app: dash.Dash
             The dash app.
         **kwargs:
@@ -197,8 +146,8 @@ class DasherLayout(ABC):
         pass
 
 
-class DasherCallback(object):
-    """ This class contains the specification of a dasher callback.
+class Callback(object):
+    """ This class contains the specification of a callback.
 
     Attributes
     ----------
@@ -212,7 +161,7 @@ class DasherCallback(object):
         The keyword arguments passed to the ``callback`` decorator.
     labels: list or dict or None
         Labels for the widgets.
-    widgets: list of DasherWidget
+    widgets: list of BaseWidget
         Generated dasher widgets for the callback.
     outputs: dash.dependencies.Output or list of dash.dependencies.Output
         Output dependencies for the callback.
@@ -238,7 +187,7 @@ class DasherCallback(object):
             The keyword arguments passed to the ``callback`` decorator.
         labels: list or dict or None
             Labels for the widgets.
-        widgets: list of DasherWidget
+        widgets: list of BaseWidget
             Generated dasher widgets for the callback.
         outputs: dash.dependencies.Output or list of dash.dependencies.Output
             Output dependencies for the callback
