+++ dasher/layout/bootstrap/__init__.py
@@ -1,2 +1,2 @@
-from .layout import DasherBootstrapLayout
+from .layout import BootstrapLayout
 from .widgets import WIDGET_SPEC
